requirements.txt
flask==2.0.1
openai==1.3.0
PyPDF2==3.0.1
python-dotenv==0.19.0
gunicorn==20.1.0
tiktoken==0.5.1 